% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_one_file_size.R
\name{get_one_file_size}
\alias{get_one_file_size}
\title{function get_one_file_size}
\usage{
get_one_file_size(filename)
}
\arguments{
\item{filename}{un nom de fichier}
}
\value{
dataframe un dateframe contenant le nom du fichier et sa taille
}
\description{
function get_one_file_size
}
\examples{
mon_dossier_temp_aleatoire <- tempfile(pattern = "toto")
dir.create(mon_dossier_temp_aleatoire)
un_fichier <- file.path(mon_dossier_temp_aleatoire, "monfichier.csv")
get_one_file_size(un_fichier)
}
